+++ montepy/__init__.py
@@ -48,6 +48,7 @@ from montepy.surface_collection import Surfaces
 from montepy.transforms import Transforms
 
 import montepy.exceptions
+import montepy.errors  # deprecated
 import sys